backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Michail Gurbanov <54814200+reqww@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
psycopg2 = "^2.8.6"
alembic = "^1.6.5"
itsdangerous = "^2.0.1"
PyJWT = "^2.1.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
emails = "^0.6"
typer = "^0.3.2"
Authlib = "^0.15.4"
httpx = "^0.18.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.4"}
aerich = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.24.1"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
